#1347 created DataPersistStgy
